Add files via upload
This program is a console-based car racing game that is written in C++ language. The game has a car that is controlled by the player, and there are enemy cars on the road that move from top to bottom. The player needs to avoid these enemy cars and score as high as possible.